Transformer working perfect :')

+++ TransformerNN/Embeddings.py
@@ -10,5 +10,5 @@ class Embeddings(nn.Module):
         self.d_model = d_model
 
     def forward(self, x):
-        #x = x.type(torch.LongTensor)
+        x = x.type(torch.LongTensor)
         return self.lut(x) * math.sqrt(self.d_model)
